Implemented partial tail call optimization for javascript translator.
Referencing a lambda using self now implements partial tail call
optimization.
It rewrites a recursive function into a loop that is exited if self is
not called. Only direct self-calls are optimised.
For now, it assumes that every function that uses self is a proper tail
recursive function. Using self with any other function will result in
undefined behaviour.